test(sidebar): refactor ts
